refactor Upload Rules
